refactor data structures
