+++ common/primitives/src/lib.rs
@@ -12,6 +12,8 @@
 	missing_docs
 )]
 
+/// Data structures for the generic use.
+pub mod ds;
 /// Structs and traits for the Messages pallet.
 pub mod messages;
 /// Structs and traits for the MSA pallet.
